Implement minimal commit log
The current implementation of commit log expects a valid Path to an
empty or unexistent directory, yet the binary needs permissions to
create it. The binary is currently only a demo, it writes to
~/loglady/ll-0 100_000 times the previous quote on println!.

The commit log is limitec to creating the dir/files and proving
interface for writing bytes to it. Next pieces to be implemented are
segment size and offset, so we can start having some fun, then
reader/indexing and such!