Rotate segments on space exhaustion
Segments must have a size, a verification of size before write, and
should error when one tries to write a bigger message size than
supported by the segment max_bytes attribute.

The CommitLog verifies the space_left bytes to ensure that the segment
will have enough space, if it doesn't it rotates to the next segment,
creating a new one.

There are many loose ends at the moment, one is that we never close the
files, this can generate an OS error for too many files open when a lot
of messages are sent and the segment size isn't big enough, it will
split into too many files.

Also, if the segment size is small enough, it will not accept messages,
e.g.: segment size of 10 bytes, but the messages are 100 bytes each, it
will loop on segment creation. We might have to add a protection, or
figure it out a way to split big messages across multiple segments.
Which, at the moment is an edge case.

The main.rs became a more of a performance overview, and the write
throughput is impressive for a single-thread sequencial file-write,
about 1M messages written in about 3 seconds.