Remove redundant reader/writer memory mappings
Since we can have "immutable" random access, given that there is no
cursor to be maintained, we don't need to keep a writer and a reader
cache.

There might be cases were, the reader/writer would make sense, if the
segment sizes were much bigger AND the reader/writer would be very far
apart in memory, AND you need to read/write at the same time, which is
not the case since we hope to keep the segment size under 100 MB, at
least for now.

This is easy to re-implement if it ever comes to that, however, it is a
unnecessary complexity right now.